+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/PeerSelectionActions.hs
@@ -15,7 +15,7 @@ module Ouroboros.Network.PeerSelection.PeerSelectionActions
 
 
 import Control.Applicative (Alternative)
-import Control.Concurrent.Class.MonadMVar (MonadMVar (..), takeMVar)
+import Control.Concurrent.Class.MonadMVar (MonadMVar (..))
 import Control.Concurrent.Class.MonadSTM.Strict
 import Control.Monad.Class.MonadAsync
 import Control.Monad.Class.MonadThrow
@@ -46,7 +46,7 @@ import Ouroboros.Network.PeerSelection.RootPeersDNS.DNSSemaphore
 import Ouroboros.Network.PeerSelection.RootPeersDNS.LocalRootPeers
 import Ouroboros.Network.PeerSelection.RootPeersDNS.PublicRootPeers
 import Ouroboros.Network.PeerSelection.State.LocalRootPeers
-import Ouroboros.Network.PeerSharing (PeerSharingController (..))
+import Ouroboros.Network.PeerSharing (PeerSharingController, requestPeers)
 import Ouroboros.Network.Protocol.PeerSharing.Type (PeerSharingAmount (..))
 import System.Random (StdGen)
 
@@ -221,9 +221,8 @@ withPeerSelectionActions
 
     requestPeerShare :: PeerSharingAmount -> peeraddr -> m (PeerSharingResult peeraddr)
     requestPeerShare amount peer = do
-      resultQueue <- newEmptyMVar
-      controller <- atomically readPeerSharingController
-      case Map.lookup peer controller of
+      controllerMap <- atomically readPeerSharingController
+      case Map.lookup peer controllerMap of
         -- Peer Registering happens asynchronously with respect to
         -- requestPeerShare. This means that there's a possible race where the
         -- Peer Selection Governor can decide to peer share request to a peer
@@ -231,7 +230,5 @@ withPeerSelectionActions
         -- going to fail, so instead of erroring we report this to the governor
         -- so it can deal with this particular case accordingly.
         Nothing -> return PeerSharingNotRegisteredYet
-        Just (PeerSharingController requestQueue) -> do
-          atomically $ putTMVar requestQueue (amount, resultQueue)
-          result <- takeMVar resultQueue
-          return (PeerSharingResult result)
+        Just psController ->
+          PeerSharingResult <$> requestPeers psController amount